refactor: create useEvaluationPanelLogic
